tests: add a @qutewm marker for bdd tests
