+++ misc/requirements/requirements-tests.txt
@@ -33,3 +33,5 @@ pytest-xvfb==1.0.0
 PyVirtualDisplay==0.2.1
 vulture==0.14
 Werkzeug==0.12.2
+python-xlib==0.17
+six==1.10.0
