+++ misc/requirements/requirements-tests.txt-raw
@@ -17,6 +17,7 @@ pytest-repeat
 pytest-rerunfailures
 pytest-travis-fold
 pytest-xvfb
+python-xlib
 vulture
 
 #@ ignore: Jinja2, MarkupSafe
