+++ tests/pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+markers =
+    qutewm: run this test with qutewm started
